inline FormattedQuestionsService implementation
