+++ src/main/java/services/FormattedQuestionsService.java
@@ -2,18 +2,26 @@ package services;
 
 import auth.HqAuth;
 import org.json.JSONArray;
+import org.json.JSONObject;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestTemplate;
 
 /**
  * Service that gets HTML formatted questions to display to the user
+ * Implemented by by requesting HQ to generate template
  */
-public interface FormattedQuestionsService {
-    QuestionResponse getFormattedQuestions(String domain, String appId, String xmlns, String instanceXml, HqAuth auth);
+public class FormattedQuestionsService {
 
-    class QuestionResponse {
+    public class QuestionResponse {
         private String formattedQuestions;
         private JSONArray questionList;
 
-        public QuestionResponse(String formattedQuestions, JSONArray questionList) {
+        private QuestionResponse(String formattedQuestions, JSONArray questionList) {
             this.formattedQuestions = formattedQuestions;
             this.questionList = questionList;
         }
@@ -24,5 +32,38 @@ public interface FormattedQuestionsService {
         public JSONArray getQuestionList() {
             return questionList;
         }
+    }
+    @Value("${commcarehq.host}")
+    private String host;
+
+    public QuestionResponse getFormattedQuestions(String domain, String appId, String xmlns, String instanceXml, HqAuth auth) {
+        RestTemplate restTemplate = new RestTemplate();
+        MultiValueMap<String, String> body = new LinkedMultiValueMap<String, String>();
+
+        body.add("instanceXml", instanceXml);
+        body.add("xmlns", xmlns);
+        body.add("appId", appId);
+
+        HttpEntity<?> entity = new HttpEntity<Object>(body, auth.getAuthHeaders());
+        ResponseEntity<String> response = restTemplate.exchange(
+                getFormattedQuestionsUrl(host, domain),
+                HttpMethod.POST,
+                entity,
+                String.class
+        );
+        if (response.getStatusCode().value() == 200) {
+            String responseBody = response.getBody();
+            JSONObject responseJSON = new JSONObject(responseBody);
+            return new QuestionResponse(
+                    responseJSON.getString("form_data"),
+                    responseJSON.getJSONArray("form_questions")
+            );
+        } else {
+            throw new RuntimeException("Error fetching debugging context");
+        }
+    }
+
+    private String getFormattedQuestionsUrl(String host, String domain) {
+        return host + "/a/" + domain + "/cloudcare/api/readable_questions/";
     }
 }
